requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
grpcio==1.67.0
grpcio-tools==1.67.0
protobuf==5.28.3
setuptools==75.2.0
sqlparse==0.5.1
tzdata==2024.2
